Crate vidformer

vidformer is a core video synthesis/transformation library. It handles the movement, control flow, and processing of video and conventional (non-video) data.

Quick links:

📦 Crates.io

📘 Documentation

🧑‍💻 Source Code

Modules

array

Arrays provide access to data.

filter

Filters create and transform video frames

service

Services provide access to media storage.

sir

Spec Intermediate Representation

source

spec

Specs declarativly define an edited video.

Structs

CodecDescriptor

Config

Config for a spesific run of a vidformer spec

Context

A context for a spec

EncoderConfig

Range

A range spesifier

Stats

Stats from a run of a vidformer spec

Enums

Error

RangeTsFormat

Select whether the output timestamps are local to the ranged segment or the entire output

Functions

codecs

List all available codecs

create_spec_hls

Create HLS artifacts for a spec

init

Instantiate FFmpeg logging

run

Execute a spec, or a range of a spec.

validate

Validate that a spec can be run.