Trait vidformer::array::Array

pub trait Array: Sync + Send {
    // Required methods
    fn domain(&self) -> Vec<Rational64>;
    fn index(&self, idx: usize) -> DataExpr;
    fn index_t(&self, idx: Rational64) -> DataExpr;
}

A trait for an array

Required Methods

fn domain(&self) -> Vec<Rational64>

Returns the domain of the array.

The domain is the set of times at which the array is defined. Each time corresponds to a single output value at that timestamp.

The output must:

Be sorted in ascending order

Contain no duplicate values

Begin with 0

fn index(&self, idx: usize) -> DataExpr

Returns the value at a given positional index.

fn index_t(&self, idx: Rational64) -> DataExpr

Returns the value at a given time index.