vidformer

Module filter

Filters create and transform video frames

A filter is a function that takes some inputs and returns a video frame. For example a filter could blur an input: MyBlur(frame, amount=5).

Modules

builtin

vidformer built-in filters

cv2

vidformer cv2 filters

Structs

Frame

A decoded video frame

FrameType

Enums

Val

A value that can be passed to a filter

Traits

Filter

A filter that can create a video frame from some inputs