vidformer::spec
Trait Spec

pub trait Spec: Sync + Send {
    // Required methods
    fn domain(&self, context: &dyn SpecContext) -> Vec<Rational64>;
    fn render(&self, context: &dyn SpecContext, t: &Rational64) -> FrameExpr;
}

A trait for a spec

A spec defines a video from a sequence of transformations. A SpecContext is provided for information about the available source videos and arrays.

Required Methods

fn domain(&self, context: &dyn SpecContext) -> Vec<Rational64>

Returns the domain of the spec.

The domain is the set of times at which the spec is defined. Each time corresponds to a single output frame at that timestamp.

The output must:

Be sorted in ascending order

Contain no duplicate values

Begin with 0

fn render(&self, context: &dyn SpecContext, t: &Rational64) -> FrameExpr

Returns the “logically” rendered frame at a given time.

This function should assume that the time is in the domain of the spec.