vidformer

Module spec

Specs declarativly define an edited video.

A spec defines an output video (as a logical time-indexed array). It creates this array by defining its timestamps (the set of times at which the spec is defined) and a function that maps each time to a frame. The frame is represented as an FrameExpr expression.

Specs are stateless and immutable. They are intended to be smaller than just storing the entire output video as an array. Additionally, this generic spec interface allows the general case of a video-editing DSL, as done in V2V, while allowing for whatever language the use case needs.

Structs

JsonSpec

A spec backed by a JSON file

Traits

Spec

A trait for a spec

SpecContext

A trait for providing information to a spec during runtime